user取得APIのリファクタ
