+++ .gitignore
@@ -1,0 +1,3 @@
+main.txt
+tmp/air.log
+tmp/main
