+++ internals/domain/model/character.go
@@ -1,0 +1,10 @@
+package model
+
+import "time"
+
+type Character struct {
+	ID        string    `json:"id"`
+	Name      string    `json:"name"`
+	CreatedAt time.Time `json:"createdAt"`
+	UpdatedAt time.Time `json:"updatedAt"`
+}
