+++ internals/domain/model/gacha.go
@@ -1,0 +1,9 @@
+package model
+
+import "time"
+
+type Gacha struct {
+	ID        string    `json:"id"`
+	CreatedAt time.Time `json:"createdAt"`
+	UpdatedAt time.Time `json:"updatedAt"`
+}
