+++ internals/domain/repository/gacha.go
@@ -1,0 +1,12 @@
+package repository
+
+// import (
+// "context"
+// "github.com/kerokerogeorge/go-gacha-api/internals/domain/model"
+
+// "github.com/gin-gonic/gin"
+// )
+
+type GachaRepository interface {
+	// FindBoardThreads(ctx context.Context, categories []string, nextDocumentId string, limit int, isOwn bool, userId string) ([]*model.BoardThread, string, error)
+}
