+++ internals/domain/repository/user.go
@@ -1,0 +1,6 @@
+package repository
+
+type UserRepository interface {
+	CreateUser(name string, token string) (string, error)
+	GetUser(token string) (string, error)
+}
