+++ internals/infrastructure/datasource/character.go
@@ -1,0 +1,1 @@
+package datasource
